Create Identity Project

+++ src/Core/Shopping.Application/Contracts/User/Models/JwtAccessTokenModel.cs
@@ -1,3 +1,3 @@
 ﻿namespace Shopping.Application.Contracts.User.Models;
 
-public record JwtAccessTokenModel(string AccessToken,int ExpirySeconds,string TokenType = "Bearer");
+public record JwtAccessTokenModel(string AccessToken,double ExpirySeconds,string TokenType = "Bearer");
